sentinela-py/pyproject.toml
[tool.poetry]
name = "sentinela"
version = "1.0.0"
description = "Vigilância total, risco zero. Sistema de monitoramento de contratos públicos"
authors = ["Sentinela Team"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlmodel = "^0.0.14"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
pyotp = "^2.9.0"
qrcode = "^7.4.2"
celery = {extras = ["redis"], version = "^5.3.6"}
redis = "^5.0.1"
boto3 = "^1.34.34"
httpx = "^0.26.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
black = "^24.1.1"
ruff = "^0.1.14"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
